+++ packages/ckb-types/index.d.ts
@@ -360,4 +360,20 @@ declare namespace CKBComponents {
     total: string
     txFee: string
   }
+
+  export interface FeeRate {
+    feeRate: Number
+  }
+
+  export type BytesOpt = Bytes | undefined
+
+  export interface WitnessArgs {
+    lock: BytesOpt // witness for lock script in input
+    inputType: BytesOpt // witness for type script in input
+    outputType: BytesOpt // witness for type script in output
+  }
+
+  export interface RawTransactionToSign extends Omit<RawTransaction, 'witnesses'> {
+    witnesses: (WitnessArgs | Witness)[]
+  }
 }